Reafactoring + aggiunta 404 page

+++ src/AdminUsersTable.js
@@ -4,12 +4,17 @@ import './mytable.css'
 import './other2_table.css'
 import AuthService from "./services/auth.service";
 import authHeader from './services/auth-header';
+import Snackbar from '@material-ui/core/Snackbar';
+import SnackbarContent from '@material-ui/core/SnackbarContent';
 
 const seller = AuthService.getCurrentUser();
 
 //const seller="angelo";
 const AdminUsersTable = (props) => {
     const [users, setUsers] = useState([])
+    const [snackColor, setSnackColor] = useState('teal');
+    const [result, setResult] = useState('');
+    const [open, setOpen] = useState(false);
     useEffect(() => {
         getData()
     },[props.trigU])
@@ -27,10 +32,15 @@ const AdminUsersTable = (props) => {
 
     //DA FARE
     const removeData = (username) => {
+       if (username === seller.username) {
+            setResult("Non puoi rimuovere te stesso...")
+            setSnackColor('orange');
+            setOpen(true);
+        } else {
         API.delete(`api/user/deleteUser/${username}`, { headers: authHeader() }).then(res => {
             const del = users.filter(users => username !== users.username)
             setUsers(del)
-        })
+        })}
     }
 
     const renderHeader = () => {
@@ -54,6 +64,12 @@ const AdminUsersTable = (props) => {
         })
     }
 
+    const handleClose = (event, reason) => {
+        if (reason === 'clickaway') {
+          return;
+        }
+        setOpen(false);
+      };
     return (
         <div>
             <table id='users'>
@@ -64,6 +80,18 @@ const AdminUsersTable = (props) => {
                     {renderBody()}
                 </tbody>
             </table>
+           <Snackbar
+        anchorOrigin={{ vertical: 'top', horizontal: 'center' }}
+        open={open}
+        autoHideDuration={3000}
+        onClose={handleClose}
+      >
+        <SnackbarContent style={{
+          backgroundColor: snackColor,
+        }}
+          message={result}
+        />
+      </Snackbar>
         </div>
     )
 }
